Update jemalloc to latest version
This is similar to https://github.com/aptos-labs/aptos-core/pull/17446. See that
PR for more details.

Similarly, we have to use a shim until everything is updated. Right now this
repo still depends on an old version of `aptos-indexer-processor-sdk`, which in
turn depends on an old version of `aptos-core`, and that pulls in
`jemalloc-sys 0.5`, which is incompatible with `tikv-jemalloc-sys 0.6`.